+++ Changelog.md
@@ -14,8 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Se agrega el modelo jetfighter a la escena (19/08/2024)
 - Se agrega rigidbody y meshcollider al modelo jetfighter, creación de prefab JetFighter (19/08/2024)
 - Se crea el script y se agrega en jetfighter, movimiento de la nave (20/08/2024)
+- Se crea el objeto Projectile y el script para el proyectil (20/08/2024)
 
 ### Changed
+
+- Se cambio el modelo jetfighter por el mismo modelo con ejes corregidos (20/08/2024)
 
 ### Deprecated